stripe/requirements.txt

pyjwt==2.3.0
pytz==2022.1
aiofiles==0.8.0
websockets==10.3
google-cloud-storage==2.3.0
asgiref==3.4.1
bcrypt==3.2.0
cffi==1.14.5
click==8.0.1
colorama==0.4.4
cryptography==3.4.7
dnspython==2.1.0
ecdsa==0.17.0
email-validator==1.1.3
fastapi==0.66.0
h11==0.12.0
idna==3.2
motor==2.4.0
passlib==1.7.4
pyasn1==0.4.8
pycparser==2.20
pydantic==1.8.2
pymongo==3.11.4
python-jose==3.3.0
python-multipart==0.0.5
rsa==4.7.2
six==1.16.0
starlette==0.14.2
typing-extensions==3.10.0.0
uvicorn==0.14.0
cachetools==5.1.0
certifi==2022.5.18.1
chardet==4.0.0
google-api-core==2.8.1
google-auth==2.6.6
google-cloud-speech==2.14.1
googleapis-common-protos==1.56.2
grpcio==1.46.3
protobuf==3.20.1
pyasn1-modules==0.2.8
requests==2.27.1
srt==3.5.2
urllib3==1.26.9
webvtt-py==0.4.6
uuid==1.30
hurry.filesize==0.9